requirements.txt
Flask==3.1.1
python-dotenv==1.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
psycopg2-binary==2.9.10
marshmallow==4.0.0
bcrypt==4.3.0